requirements.txt
annotated-types==0.7.0
cachecontrol==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
comtypes==1.4.7
cryptography==43.0.1
filelock==3.16.1
firebase-admin==6.5.0
fsspec==2024.9.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-cloud-texttospeech==2.17.2
google-cloud-translate==3.16.0
google-cloud-vision==3.7.4
google-crc32c==1.6.0
google-generativeai==0.8.2
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
numpy==2.0.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pvporcupine==3.0.3
pyasn1==0.6.1
pyasn1-modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pygame==2.6.1
PyJWT==2.9.0
pyparsing==3.1.4
pypiwin32==223
pytesseract==0.3.13
pyttsx3==2.98
pywin32==306
requests==2.32.3
rsa==4.9
SpeechRecognition==3.10.4
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
